requirements.txt
langchain
langchain_community
langchain-google-genai
streamlit
python-dotenv